Join db and tsv column classes, and extract storage module
