+++ openverse_catalog/dags/provider_api_scripts/europeana.py
@@ -16,7 +16,7 @@ from datetime import datetime, timedelta, timezone
 
 from common.licenses.licenses import get_license_info
 from common.requester import DelayedRequester
-from common.storage.image import ImageStore
+from storage.image import ImageStore
 from util.loader import provider_details as prov
