+++ openverse_catalog/dags/provider_api_scripts/flickr.py
@@ -18,7 +18,7 @@ from datetime import datetime, timedelta, timezone
 import lxml.html as html
 from common.licenses.licenses import get_license_info
 from common.requester import DelayedRequester
-from common.storage.image import ImageStore
+from storage.image import ImageStore
 from util.loader import provider_details as prov
