+++ openverse_catalog/dags/storage/test_audio.py
@@ -2,8 +2,9 @@ import logging
 from unittest.mock import patch
 
 import pytest
+import storage.columns
 from common.licenses.licenses import LicenseInfo
-from common.storage import audio
+from storage import audio
 
 
 logging.basicConfig(
@@ -219,10 +220,11 @@ def test_create_tsv_row_creates_alt_files(
         return value
 
     with patch.object(
-        audio.columns.urls, "validate_url_string", side_effect=mock_url_validator
+        storage.columns.urls,
+        "validate_url_string",
+        side_effect=mock_url_validator,
     ):
         actual_row = audio_store._create_tsv_row(test_audio)
-        print(f"\nActual row: {actual_row}")
         expected_row = (
             "\t".join(
                 [
@@ -277,7 +279,7 @@ def test_create_tsv_row_creates_audio_set(
         return value
 
     with patch.object(
-        audio.columns.urls, "validate_url_string", side_effect=mock_url_validator
+        storage.columns.urls, "validate_url_string", side_effect=mock_url_validator
     ):
         actual_row = audio_store._create_tsv_row(test_audio)
         expected_row = (
@@ -399,7 +401,7 @@ def test_create_tsv_row_properly_places_entries(monkeypatch):
     def mock_validate_url(url_string):
         return url_string
 
-    monkeypatch.setattr(audio.columns.urls, "validate_url_string", mock_validate_url)
+    monkeypatch.setattr(storage.columns.urls, "validate_url_string", mock_validate_url)
     audio_store = audio.AudioStore()
     req_args_dict = {
         "foreign_landing_url": "https://landing_page.com",
