+++ openverse_catalog/dags/storage/test_columns.py
@@ -2,7 +2,8 @@ import logging
 import string
 
 import tldextract
-from common.storage import columns
+from storage import columns
+from storage.columns import Datatype
 
 
 logging.basicConfig(
@@ -17,7 +18,7 @@ class TruncateColumn(columns.Column):
     def __init__(self, size, truncate):
         self.SIZE = size
         self.TRUNCATE = truncate
-        super().__init__("test_column", False)
+        super().__init__("test_column", False, datatype=Datatype.char)
 
     def prepare_string(self, value):
         return self._Column__enforce_char_limit(value, self.SIZE, self.TRUNCATE)
