+++ openverse_catalog/dags/storage/test_image.py
@@ -1,13 +1,9 @@
-"""
-MediaStore is an abstract class, so to test it we
-use one of the inheriting classes, ImageStore
-"""
 import logging
-from unittest.mock import patch
 
 import pytest
-from common.licenses.licenses import LicenseInfo, get_license_info
-from common.storage import image
+from common.licenses.licenses import LicenseInfo
+from storage import image, util
+from storage.columns import urls
 
 
 logging.basicConfig(
@@ -16,7 +12,6 @@ logging.basicConfig(
 
 logger = logging.getLogger(__name__)
 
-IMAGE_COLUMN_NAMES = [x.NAME for x in image.IMAGE_TSV_COLUMNS]
 
 PD_LICENSE_INFO = LicenseInfo(
     "zero", "1.0", "https://creativecommons.org/publicdomain/zero/1.0/", None
@@ -31,7 +26,7 @@ def setup_env(monkeypatch):
     monkeypatch.setenv("OUTPUT_DIR", "/tmp")
 
 
-def test_MediaStore_uses_OUTPUT_DIR_variable(
+def test_ImageStore_uses_OUTPUT_DIR_variable(
     monkeypatch,
 ):
     testing_output_dir = "/my_output_dir"
@@ -40,7 +35,7 @@ def test_MediaStore_uses_OUTPUT_DIR_variable(
     assert testing_output_dir in image_store._OUTPUT_PATH
 
 
-def test_MediaStore_falls_back_to_tmp_output_dir_variable(
+def test_ImageStore_falls_back_to_tmp_output_dir_variable(
     monkeypatch,
     setup_env,
 ):
@@ -49,19 +44,15 @@ def test_MediaStore_falls_back_to_tmp_output_dir_variable(
     assert "/tmp" in image_store._OUTPUT_PATH
 
 
-def test_MediaStore_includes_provider_in_output_file_string():
+def test_ImageStore_includes_provider_in_output_file_string(
+    setup_env,
+):
     image_store = image.ImageStore("test_provider")
     assert type(image_store._OUTPUT_PATH) == str
     assert "test_provider" in image_store._OUTPUT_PATH
 
 
-def test_MediaStore_includes_media_type_in_output_file_string():
-    image_store = image.ImageStore("test_provider")
-    assert type(image_store._OUTPUT_PATH) == str
-    assert "image" in image_store._OUTPUT_PATH
-
-
-def test_MediaStore_add_item_adds_realistic_image_to_buffer():
+def test_ImageStore_add_item_adds_realistic_image_to_buffer(setup_env):
     image_store = image.ImageStore(provider="testing_provider")
     image_store.add_item(
         foreign_landing_url="https://images.org/image01",
@@ -71,7 +62,9 @@ def test_MediaStore_add_item_adds_realistic_image_to_buffer():
     assert len(image_store._media_buffer) == 1
 
 
-def test_MediaStore_add_item_adds_multiple_images_to_buffer():
+def test_ImageStore_add_item_adds_multiple_images_to_buffer(
+    setup_env,
+):
     image_store = image.ImageStore(provider="testing_provider")
     image_store.add_item(
         foreign_landing_url="https://images.org/image01",
@@ -96,7 +89,10 @@ def test_MediaStore_add_item_adds_multiple_images_to_buffer():
     assert len(image_store._media_buffer) == 4
 
 
-def test_MediaStore_add_item_flushes_buffer(tmpdir):
+def test_ImageStore_add_item_flushes_buffer(
+    setup_env,
+    tmpdir,
+):
     output_file = "testing.tsv"
     tmp_directory = tmpdir
     output_dir = str(tmp_directory)
@@ -135,12 +131,12 @@ def test_MediaStore_add_item_flushes_buffer(tmpdir):
     assert len(lines) == 4  # recall the last '\n' will create an empty line.
 
 
-def test_MediaStore_commit_writes_nothing_if_no_lines_in_buffer():
+def test_ImageStore_commit_writes_nothing_if_no_lines_in_buffer():
     image_store = image.ImageStore(output_dir="/path/does/not/exist")
     image_store.commit()
 
 
-def test_MediaStore_produces_correct_total_images():
+def test_ImageStore_produces_correct_total_images(setup_env):
     image_store = image.ImageStore(provider="testing_provider")
     image_store.add_item(
         foreign_landing_url="https://images.org/image01",
@@ -157,143 +153,66 @@ def test_MediaStore_produces_correct_total_images():
         image_url="https://images.org/image03.jpg",
         license_info=PD_LICENSE_INFO,
     )
+    image_store.add_item(
+        foreign_landing_url="https://images.org/image04",
+        image_url="https://images.org/image04.jpg",
+        license_info=LicenseInfo(None, None, None, None),
+    )
     assert image_store.total_items == 3
 
 
-def test_MediaStore_clean_media_metadata_does_not_change_required_media_arguments(
+def test_ImageStore_get_image_places_given_args(
     monkeypatch,
 ):
-    image_url = "test_url"
-    foreign_landing_url = "foreign_landing_url"
-    image_store = image.ImageStore()
-    image_data = {
+    image_store = image.ImageStore(provider="testing_provider")
+    args_dict = {
+        "foreign_landing_url": "https://landing_page.com",
+        "image_url": "https://imageurl.com",
         "license_info": BY_LICENSE_INFO,
-        "foreign_landing_url": foreign_landing_url,
-        "image_url": image_url,
-        "thumbnail_url": None,
-        "foreign_identifier": None,
-    }
-    cleaned_data = image_store.clean_media_metadata(**image_data)
-
-    assert cleaned_data["image_url"] == image_url
-    assert cleaned_data["foreign_landing_url"] == foreign_landing_url
-
-
-def test_MediaStore_clean_media_metadata_adds_provider(
-    monkeypatch,
-):
-    provider = "test_provider"
-    image_store = image.ImageStore(provider=provider)
-    image_data = {
-        "license_info": BY_LICENSE_INFO,
-        "foreign_landing_url": None,
-        "image_url": None,
-    }
-    cleaned_data = image_store.clean_media_metadata(**image_data)
-
-    assert cleaned_data["provider"] == provider
-
-
-def test_MediaStore_clean_media_metadata_adds_filesize(
-    monkeypatch,
-):
-    image_store = image.ImageStore()
-    image_data = {
-        "license_info": BY_LICENSE_INFO,
-    }
-    cleaned_data = image_store.clean_media_metadata(**image_data)
-
-    assert "filesize" in cleaned_data
-    assert cleaned_data["filesize"] is None
-
-
-def test_MediaStore_clean_media_metadata_removes_license_urls(
-    monkeypatch,
-):
-    image_store = image.ImageStore()
-    image_data = {
-        "license_info": BY_LICENSE_INFO,
-        "foreign_landing_url": None,
-        "image_url": None,
-        "thumbnail_url": None,
-        "foreign_identifier": None,
-    }
-    cleaned_data = image_store.clean_media_metadata(**image_data)
-
-    assert "license_url" not in cleaned_data
-    assert "raw_license_url" not in cleaned_data
-
-
-def test_MediaStore_clean_media_metadata_replaces_license_url_with_license_info(
-    monkeypatch,
-):
-    image_store = image.ImageStore()
-    image_data = {
-        "license_info": BY_LICENSE_INFO,
-    }
-    cleaned_data = image_store.clean_media_metadata(**image_data)
-
-    expected_license = "by"
-    expected_version = "4.0"
-    assert cleaned_data["license_"] == expected_license
-    assert cleaned_data["license_version"] == expected_version
-    assert "license_url" not in cleaned_data
-
-
-def test_MediaStore_clean_media_metadata_adds_license_urls_to_meta_data(
-    monkeypatch,
-):
-    raw_license_url = "raw_license"
-    license_url = "https://creativecommons.org/licenses/by-nc-nd/4.0/"
-    image_store = image.ImageStore()
-    image_data = {
-        "license_info": LicenseInfo(
-            "by-nc-nd",
-            "4.0",
-            license_url,
-            raw_license_url,
-        ),
-        "foreign_landing_url": None,
-        "image_url": None,
-        "thumbnail_url": None,
-        "foreign_identifier": None,
+        "foreign_identifier": "foreign_id",
+        "thumbnail_url": "https://thumbnail.com",
+        "width": 200,
+        "height": 500,
+        "creator": "tyler",
+        "creator_url": "https://creatorurl.com",
+        "title": "agreatpicture",
+        "meta_data": {"description": "cat picture"},
+        "raw_tags": [{"name": "tag1", "provider": "testing"}],
+        "watermarked": "f",
+        "source": "testing_source",
         "ingestion_type": "provider_api",
     }
-    cleaned_data = image_store.clean_media_metadata(**image_data)
 
-    assert cleaned_data["meta_data"]["license_url"] == license_url
-    assert cleaned_data["meta_data"]["raw_license_url"] == raw_license_url
+    def mock_get_source(source, provider):
+        return source
+
+    monkeypatch.setattr(util, "get_source", mock_get_source)
+
+    def mock_enrich_tags(tags):
+        return tags
+
+    monkeypatch.setattr(image_store, "_enrich_tags", mock_enrich_tags)
+
+    actual_image = image_store._get_image(**args_dict)
+    args_dict["tags"] = args_dict.pop("raw_tags")
+    args_dict["provider"] = "testing_provider"
+    args_dict["filesize"] = None
+    args_dict["license_"] = args_dict.get("license_info").license
+    args_dict["license_version"] = args_dict.pop("license_info").version
+
+    assert actual_image == image.Image(**args_dict)
 
 
-def test_MediaStore_get_image_gets_source(
+def test_ImageStore_get_media_gets_source(
     monkeypatch,
+    setup_env,
 ):
     image_store = image.ImageStore()
 
-    actual_image = image_store._get_image(
-        license_info=BY_LICENSE_INFO,
-        foreign_landing_url=None,
-        image_url=None,
-        thumbnail_url=None,
-        foreign_identifier=None,
-        width=None,
-        height=None,
-        creator=None,
-        creator_url=None,
-        title=None,
-        meta_data=None,
-        raw_tags=None,
-        watermarked=None,
-        source="diff_source",
-        ingestion_type=None,
-    )
-    assert actual_image.source == "diff_source"
+    def mock_get_source(source, provider):
+        return "diff_source"
 
-
-def test_MediaStore_sets_source_to_provider_if_source_is_none(
-    monkeypatch,
-):
-    image_store = image.ImageStore(provider="test_provider")
+    monkeypatch.setattr(util, "get_source", mock_get_source)
 
     actual_image = image_store._get_image(
         license_info=BY_LICENSE_INFO,
@@ -312,147 +231,48 @@ def test_MediaStore_sets_source_to_provider_if_source_is_none(
         source=None,
         ingestion_type=None,
     )
-    assert actual_image.source == "test_provider"
+    assert actual_image.source == "diff_source"
 
 
-def test_MediaStore_add_image_replaces_non_dict_meta_data_with_no_license_url():
+def test_ImageStore_get_image_creates_meta_data_with_valid_license_url():
+    license_url = "https://my.license.url"
     image_store = image.ImageStore()
+    licence_info = LicenseInfo(
+        "by", "4.0", "https://creativecommons.org/licenses/by/4.0/", license_url
+    )
 
-    def item_saver(arg):
-        pass
-
-    with patch.object(image_store, "save_item", side_effect=item_saver) as mock_save:
-        image_store.add_item(
-            license_info=BY_LICENSE_INFO,
-            foreign_landing_url="",
-            image_url="",
-            thumbnail_url=None,
-            foreign_identifier=None,
-            width=None,
-            height=None,
-            creator=None,
-            creator_url=None,
-            title=None,
-            meta_data="notadict",
-            raw_tags=None,
-            watermarked=None,
-            source=None,
-            ingestion_type=None,
-        )
-    actual_image = mock_save.call_args[0][0]
+    actual_image = image_store._get_image(
+        license_info=licence_info,
+        foreign_landing_url=None,
+        image_url=None,
+        thumbnail_url=None,
+        foreign_identifier=None,
+        width=None,
+        height=None,
+        creator=None,
+        creator_url=None,
+        title=None,
+        meta_data=None,
+        raw_tags=None,
+        watermarked=None,
+        source=None,
+        ingestion_type=None,
+    )
     assert actual_image.meta_data == {
         "license_url": "https://creativecommons.org/licenses/by/4.0/",
-        "raw_license_url": None,
+        "raw_license_url": license_url,
     }
 
 
-def test_MediaStore_add_item_creates_meta_data_with_valid_license_url(
-    monkeypatch, setup_env
-):
-    image_store = image.ImageStore()
-
-    license_url = "https://my.license.url"
-    valid_license_url = "https://creativecommons.org/licenses/by/4.0/"
-
-    def item_saver(arg):
-        pass
-
-    with patch.object(image_store, "save_item", side_effect=item_saver) as mock_save:
-        image_store.add_item(
-            license_info=LicenseInfo("by", "4.0", valid_license_url, license_url),
-            foreign_landing_url="",
-            image_url="",
-            thumbnail_url=None,
-            foreign_identifier=None,
-            width=None,
-            height=None,
-            creator=None,
-            creator_url=None,
-            title=None,
-            meta_data=None,
-            raw_tags=None,
-            watermarked=None,
-            source=None,
-            ingestion_type=None,
-        )
-        actual_image = mock_save.call_args[0][0]
-
-        assert actual_image.meta_data == {
-            "license_url": valid_license_url,
-            "raw_license_url": license_url,
-        }
-
-
-def test_MediaStore_add_item_adds_valid_license_url_to_dict_meta_data(
-    monkeypatch, setup_env
-):
-    image_store = image.ImageStore()
-
-    license_url = "https://my.license.url"
-    valid_license_url = "https://creativecommons.org/licenses/by/4.0/"
-
-    def item_saver(arg):
-        pass
-
-    with patch.object(image_store, "save_item", side_effect=item_saver) as mock_save:
-        image_store.add_item(
-            license_info=LicenseInfo("by", "4.0", valid_license_url, license_url),
-            foreign_landing_url="",
-            image_url="",
-            thumbnail_url=None,
-            foreign_identifier=None,
-            width=None,
-            height=None,
-            creator=None,
-            creator_url=None,
-            title=None,
-            meta_data={"key1": "val1"},
-            raw_tags=None,
-            watermarked=None,
-            source=None,
-            ingestion_type=None,
-        )
-        actual_image = mock_save.call_args[0][0]
-
-        assert actual_image.meta_data == {
-            "key1": "val1",
-            "license_url": valid_license_url,
-            "raw_license_url": license_url,
-        }
-
-
-def test_ImageStore_add_item_fixes_invalid_license_url():
-    image_store = image.ImageStore()
-
-    original_url = "https://license/url"
-    updated_url = "https://creativecommons.org/licenses/by-nc-sa/2.0/"
-
-    def item_saver(arg):
-        pass
-
-    with patch.object(image_store, "save_item", side_effect=item_saver) as mock_save:
-        image_store.add_item(
-            license_info=LicenseInfo("by-nc-sa", "2.0", updated_url, original_url),
-            foreign_landing_url="",
-            image_url="",
-            meta_data={},
-        )
-    actual_image = mock_save.call_args[0][0]
-
-    assert actual_image.meta_data == {
-        "license_url": updated_url,
-        "raw_license_url": original_url,
-    }
-
-
-def test_MediaStore_get_image_enriches_singleton_tags():
+def test_ImageStore_get_image_enriches_singleton_tags():
     image_store = image.ImageStore("test_provider")
 
     actual_image = image_store._get_image(
-        license_info=get_license_info(
-            license_="by-sa",
-            license_version="4.0",
-            license_url="https://license/url",
+        license_info=LicenseInfo(
+            "by-nc-nd",
+            "4.0",
+            "https://creativecommons.org/licenses/by-nc-nd/4.0/",
+            "https://license/url",
         ),
         foreign_landing_url=None,
         image_url=None,
@@ -467,13 +287,13 @@ def test_MediaStore_get_image_enriches_singleton_tags():
         raw_tags=["lone"],
         watermarked=None,
         source=None,
-        ingestion_type=None,
+        ingestion_type="provider_api",
     )
 
     assert actual_image.tags == [{"name": "lone", "provider": "test_provider"}]
 
 
-def test_MediaStore_get_image_tag_blacklist():
+def test_ImageStore_get_image_tag_blacklist():
     raw_tags = [
         "cc0",
         "valid",
@@ -485,36 +305,32 @@ def test_MediaStore_get_image_tag_blacklist():
     image_store = image.ImageStore("test_provider")
 
     actual_image = image_store._get_image(
-        license_info=get_license_info(
-            license_="by",
-            license_version="4.0",
-        ),
+        license_info=BY_LICENSE_INFO,
         foreign_landing_url=None,
         image_url=None,
-        meta_data=None,
-        raw_tags=raw_tags,
-        foreign_identifier=None,
         thumbnail_url=None,
+        foreign_identifier=None,
         width=None,
         height=None,
         creator=None,
         creator_url=None,
         title=None,
+        meta_data=None,
+        raw_tags=raw_tags,
         watermarked=None,
-        ingestion_type=None,
+        source=None,
+        ingestion_type="provider_api",
     )
 
     assert actual_image.tags == [{"name": "valid", "provider": "test_provider"}]
 
 
-def test_MediaStore_get_image_enriches_multiple_tags():
+def test_ImageStore_get_image_enriches_multiple_tags(
+    setup_env,
+):
     image_store = image.ImageStore("test_provider")
     actual_image = image_store._get_image(
-        license_info=get_license_info(
-            license_url="https://license/url",
-            license_="by",
-            license_version="4.0",
-        ),
+        license_info=BY_LICENSE_INFO,
         foreign_landing_url=None,
         image_url=None,
         thumbnail_url=None,
@@ -528,7 +344,7 @@ def test_MediaStore_get_image_enriches_multiple_tags():
         raw_tags=["tagone", "tag2", "tag3"],
         watermarked=None,
         source=None,
-        ingestion_type=None,
+        ingestion_type="provider_api",
     )
 
     assert actual_image.tags == [
@@ -547,11 +363,7 @@ def test_ImageStore_get_image_leaves_preenriched_tags(setup_env):
     ]
 
     actual_image = image_store._get_image(
-        license_info=get_license_info(
-            license_url="https://license/url",
-            license_="by",
-            license_version="4.0",
-        ),
+        license_info=BY_LICENSE_INFO,
         foreign_landing_url=None,
         image_url=None,
         thumbnail_url=None,
@@ -565,7 +377,7 @@ def test_ImageStore_get_image_leaves_preenriched_tags(setup_env):
         raw_tags=tags,
         watermarked=None,
         source=None,
-        ingestion_type=None,
+        ingestion_type="provider_api",
     )
 
     assert actual_image.tags == tags
@@ -576,11 +388,7 @@ def test_ImageStore_get_image_nones_nonlist_tags():
     tags = "notalist"
 
     actual_image = image_store._get_image(
-        license_info=get_license_info(
-            license_url="https://license/url",
-            license_="by",
-            license_version="4.0",
-        ),
+        license_info=BY_LICENSE_INFO,
         foreign_landing_url=None,
         image_url=None,
         thumbnail_url=None,
@@ -594,14 +402,16 @@ def test_ImageStore_get_image_nones_nonlist_tags():
         raw_tags=tags,
         watermarked=None,
         source=None,
-        ingestion_type=None,
+        ingestion_type="provider_api",
     )
 
     assert actual_image.tags is None
 
 
 @pytest.fixture
-def default_image_args():
+def default_image_args(
+    setup_env,
+):
     return dict(
         foreign_identifier=None,
         foreign_landing_url="https://image.org",
@@ -620,12 +430,13 @@ def default_image_args():
         watermarked=None,
         provider=None,
         source=None,
-        ingestion_type=None,
+        ingestion_type="provider_api",
     )
 
 
 def test_create_tsv_row_non_none_if_req_fields(
     default_image_args,
+    setup_env,
 ):
     image_store = image.ImageStore()
     test_image = image.Image(**default_image_args)
@@ -635,6 +446,7 @@ def test_create_tsv_row_non_none_if_req_fields(
 
 def test_create_tsv_row_none_if_no_foreign_landing_url(
     default_image_args,
+    setup_env,
 ):
     image_store = image.ImageStore()
     image_args = default_image_args
@@ -647,6 +459,7 @@ def test_create_tsv_row_none_if_no_foreign_landing_url(
 
 def test_create_tsv_row_none_if_no_license(
     default_image_args,
+    setup_env,
 ):
     image_store = image.ImageStore()
     image_args = default_image_args
@@ -659,6 +472,7 @@ def test_create_tsv_row_none_if_no_license(
 
 def test_create_tsv_row_none_if_no_license_version(
     default_image_args,
+    setup_env,
 ):
     image_store = image.ImageStore()
     image_args = default_image_args
@@ -671,6 +485,7 @@ def test_create_tsv_row_none_if_no_license_version(
 
 def test_create_tsv_row_returns_none_if_missing_image_url(
     default_image_args,
+    setup_env,
 ):
     image_store = image.ImageStore()
     image_args = default_image_args
@@ -683,6 +498,7 @@ def test_create_tsv_row_returns_none_if_missing_image_url(
 
 def test_create_tsv_row_handles_empty_dict_and_tags(
     default_image_args,
+    setup_env,
 ):
     image_store = image.ImageStore()
     meta_data = {}
@@ -693,12 +509,7 @@ def test_create_tsv_row_handles_empty_dict_and_tags(
     test_image = image.Image(**image_args)
 
     actual_row = image_store._create_tsv_row(test_image).split("\t")
-    meta_data_col_id = IMAGE_COLUMN_NAMES.index("meta_data")
-    tags_col_id = IMAGE_COLUMN_NAMES.index("tags")
-    actual_meta_data, actual_tags = (
-        actual_row[meta_data_col_id],
-        actual_row[tags_col_id],
-    )
+    actual_meta_data, actual_tags = actual_row[12], actual_row[13]
     expect_meta_data, expect_tags = "\\N", "\\N"
     assert expect_meta_data == actual_meta_data
     assert expect_tags == actual_tags
@@ -706,44 +517,43 @@ def test_create_tsv_row_handles_empty_dict_and_tags(
 
 def test_create_tsv_row_turns_empty_into_nullchar(
     default_image_args,
+    setup_env,
 ):
-    """
-    Null values are converted into `N/A` in tsv files
-    This test first selects all the media properties with value None,
-    and then checks if all corresponding tsv values are `N/A`.
-    The last element has a new line at the end, so we check it separately
-    """
     image_store = image.ImageStore()
     image_args = default_image_args
+    image_args["ingestion_type"] = None
     test_image = image.Image(**image_args)
 
-    none_fields = [
-        i for i, x in enumerate(test_image._fields) if getattr(test_image, x) is None
-    ]
-    # none_field_names = [test_image._fields[x] for x in none_fields]
-
     actual_row = image_store._create_tsv_row(test_image).split("\t")
+    assert (
+        all(
+            [
+                actual_row[i] == "\\N"
+                for i in [0, 3, 4, 5, 6, 9, 10, 11, 12, 13, 14, 15, 16]
+            ]
+        )
+        is True
+    )
     assert actual_row[-1] == "\\N\n"
 
-    actual_row[-1] = "\\N"
-    assert all([actual_row[i] == "\\N" for i in none_fields]) is True
 
-
-def test_create_tsv_row_properly_places_entries(monkeypatch):
+def test_create_tsv_row_properly_places_entries(setup_env, monkeypatch):
     def mock_validate_url(url_string):
         return url_string
 
-    monkeypatch.setattr(image.columns.urls, "validate_url_string", mock_validate_url)
+    monkeypatch.setattr(urls, "validate_url_string", mock_validate_url)
     image_store = image.ImageStore()
     req_args_dict = {
         "foreign_landing_url": "https://landing_page.com",
-        "image_url": "https://imageurl.com",
+        "image_url": "http://imageurl.com",
         "license_": "testlicense",
         "license_version": "1.0",
     }
     args_dict = {
         "foreign_identifier": "foreign_id",
-        "thumbnail_url": "https://thumbnail.com",
+        "thumbnail_url": "http://thumbnail.com",
+        "width": 200,
+        "height": 500,
         "filesize": None,
         "creator": "tyler",
         "creator_url": "https://creatorurl.com",
@@ -753,9 +563,7 @@ def test_create_tsv_row_properly_places_entries(monkeypatch):
         "watermarked": "f",
         "provider": "testing_provider",
         "source": "testing_source",
-        "ingestion_type": "testing_ingestion",
-        "width": 200,
-        "height": 500,
+        "ingestion_type": "provider_api",
     }
     args_dict.update(req_args_dict)
 
@@ -766,8 +574,10 @@ def test_create_tsv_row_properly_places_entries(monkeypatch):
             [
                 "foreign_id",
                 "https://landing_page.com",
-                "https://imageurl.com",
-                "https://thumbnail.com",
+                "http://imageurl.com",
+                "http://thumbnail.com",
+                "200",
+                "500",
                 "\\N",
                 "testlicense",
                 "1.0",
@@ -779,9 +589,7 @@ def test_create_tsv_row_properly_places_entries(monkeypatch):
                 "f",
                 "testing_provider",
                 "testing_source",
-                "testing_ingestion",
-                "200",
-                "500",
+                "provider_api",
             ]
         )
         + "\n"
