+++ openverse_catalog/dags/util/loader/loader.py
@@ -2,23 +2,21 @@ from util.loader import paths, s3, sql
 from util.loader.paths import _extract_media_type
 
 
-def load_local_data(output_dir, postgres_conn_id, identifier, overwrite=False):
+def load_local_data(output_dir, postgres_conn_id, identifier, ti, overwrite=False):
     tsv_file_name = paths.get_staged_file(output_dir, identifier)
+    tsv_version = ti.xcom_pull(task_ids="stage_oldest_tsv_file", key="tsv_version")
+
     media_type = _extract_media_type(tsv_file_name)
     sql.load_local_data_to_intermediate_table(
         postgres_conn_id, tsv_file_name, identifier
     )
     if overwrite is True:
         sql.overwrite_records_in_db_table(
-            postgres_conn_id,
-            identifier,
-            media_type=media_type,
+            postgres_conn_id, identifier, media_type=media_type, tsv_version=tsv_version
         )
     else:
         sql.upsert_records_to_db_table(
-            postgres_conn_id,
-            identifier,
-            media_type=media_type,
+            postgres_conn_id, identifier, media_type=media_type, tsv_version=tsv_version
         )
 
 
@@ -39,6 +37,8 @@ def load_s3_data(
     overwrite=False,
 ):
     media_type = ti.xcom_pull(task_ids="stage_oldest_tsv_file", key="media_type")
+    tsv_version = ti.xcom_pull(task_ids="stage_oldest_tsv_file", key="tsv_version")
+
     if media_type is None:
         media_type = "image"
     tsv_key = s3.get_staged_s3_object(
@@ -47,9 +47,12 @@ def load_s3_data(
     sql.load_s3_data_to_intermediate_table(
         postgres_conn_id, bucket, tsv_key, identifier, media_type
     )
+
     if overwrite is True:
-        sql.overwrite_records_in_db_table(postgres_conn_id, identifier, media_type)
+        sql.overwrite_records_in_db_table(
+            postgres_conn_id, identifier, media_type=media_type, tsv_version=tsv_version
+        )
     else:
         sql.upsert_records_to_db_table(
-            postgres_conn_id, identifier, media_type=media_type
+            postgres_conn_id, identifier, media_type=media_type, tsv_version=tsv_version
         )
