+++ openverse_catalog/dags/util/loader/test_resources/foreign_identifier_missing.tsv
@@ -1,2 +1,2 @@
-7789139	https://www.thingiverse.com/thing:2823006	https://cdn.thingiverse.com/renders/cd/2f/f4/ba/85/1892c3b6344dc168a18897359a0c97a9_display_large.jpg	https://cdn.thingiverse.com/renders/cd/2f/f4/ba/85/1892c3b6344dc168a18897359a0c97a9_display_medium.jpg	\N	cc0	1.0	Walter Hsiao	https://www.thingiverse.com/walter	Air Spinner	{"description": "*Update 2018-04-21: Added two additional versions* I had no idea what this was for when I uploaded it. I didn't think it would be useful for anything other than as a 3D print demo, I just made it to satisfy my curiosity. But if you check out comments below, [guttyr](https://www.thingiverse.com/guttyr/about) discovered you can spin it up by blowing on it, and it can be used by speech therapists to help special needs kids get better at breath support and control. Different parts of the spinner will spin depending where you blow on it, I'm still figuring which locations and angles work best. Maybe it'll help me too as I'm frequently out of breath when I'm at altitude or underwater. It can be printed with a 0.5mm extrusion width and 2 perimeters without infill. Printed in [Atomic Filament Granite Grey PLA](https://atomicfilament.com/products/gray-granite-pla-filament) and [Octave Precision Gold PLA*](http://amzn.to/2p7JPU0) New versions printed in [Atomic Filament Starry Night PLA](https://atomicfilament.com/collections/opaque-pla-filaments-1/products/translucent-starry-night-pla) and [3D Solutech Transparent Blue PETG*](https://amzn.to/2K3migu) **affiliate links*", "3d_model": "https://cdn.thingiverse.com/assets/49/2e/e5/2b/26/Gimball_Test_Print.stl"}	[{"name": "air", "provider": "thingiverse"}, {"name": "Demo", "provider": "thingiverse"}, {"name": "Gimbal", "provider": "thingiverse"}, {"name": "Spinner", "provider": "thingiverse"}, {"name": "Test", "provider": "thingiverse"}, {"name": "Useless", "provider": "thingiverse"}]	f	thingiverse	thingiverse	provider_api	\N	\N
-\N	https://www.thingiverse.com/thing:2823006	https://cdn.thingiverse.com/renders/ed/47/66/f7/43/3dee6caf0e321f796b4d34b78945a4ee_display_large.jpg	https://cdn.thingiverse.com/renders/ed/47/66/f7/43/3dee6caf0e321f796b4d34b78945a4ee_display_medium.jpg	\N	cc0	1.0	Walter Hsiao	https://www.thingiverse.com/walter	Air Spinner	{"description": "*Update 2018-04-21: Added two additional versions* I had no idea what this was for when I uploaded it. I didn't think it would be useful for anything other than as a 3D print demo, I just made it to satisfy my curiosity. But if you check out comments below, [guttyr](https://www.thingiverse.com/guttyr/about) discovered you can spin it up by blowing on it, and it can be used by speech therapists to help special needs kids get better at breath support and control. Different parts of the spinner will spin depending where you blow on it, I'm still figuring which locations and angles work best. Maybe it'll help me too as I'm frequently out of breath when I'm at altitude or underwater. It can be printed with a 0.5mm extrusion width and 2 perimeters without infill. Printed in [Atomic Filament Granite Grey PLA](https://atomicfilament.com/products/gray-granite-pla-filament) and [Octave Precision Gold PLA*](http://amzn.to/2p7JPU0) New versions printed in [Atomic Filament Starry Night PLA](https://atomicfilament.com/collections/opaque-pla-filaments-1/products/translucent-starry-night-pla) and [3D Solutech Transparent Blue PETG*](https://amzn.to/2K3migu) **affiliate links*", "3d_model": "https://cdn.thingiverse.com/assets/c7/e9/4a/9f/6b/Air_Spinner_2_-_Solid.stl"}	[{"name": "air", "provider": "thingiverse"}, {"name": "Demo", "provider": "thingiverse"}, {"name": "Gimbal", "provider": "thingiverse"}, {"name": "Spinner", "provider": "thingiverse"}, {"name": "Test", "provider": "thingiverse"}, {"name": "Useless", "provider": "thingiverse"}]	f	thingiverse	thingiverse	provider_api	\N	\N
+7789139	https://www.thingiverse.com/thing:2823006	https://cdn.thingiverse.com/renders/cd/2f/f4/ba/85/1892c3b6344dc168a18897359a0c97a9_display_large.jpg	https://cdn.thingiverse.com/renders/cd/2f/f4/ba/85/1892c3b6344dc168a18897359a0c97a9_display_medium.jpg	\N	\N	\N	cc0	1.0	Walter Hsiao	https://www.thingiverse.com/walter	Air Spinner	{"description": "*Update 2018-04-21: Added two additional versions* I had no idea what this was for when I uploaded it. I didn't think it would be useful for anything other than as a 3D print demo, I just made it to satisfy my curiosity. But if you check out comments below, [guttyr](https://www.thingiverse.com/guttyr/about) discovered you can spin it up by blowing on it, and it can be used by speech therapists to help special needs kids get better at breath support and control. Different parts of the spinner will spin depending where you blow on it, I'm still figuring which locations and angles work best. Maybe it'll help me too as I'm frequently out of breath when I'm at altitude or underwater. It can be printed with a 0.5mm extrusion width and 2 perimeters without infill. Printed in [Atomic Filament Granite Grey PLA](https://atomicfilament.com/products/gray-granite-pla-filament) and [Octave Precision Gold PLA*](http://amzn.to/2p7JPU0) New versions printed in [Atomic Filament Starry Night PLA](https://atomicfilament.com/collections/opaque-pla-filaments-1/products/translucent-starry-night-pla) and [3D Solutech Transparent Blue PETG*](https://amzn.to/2K3migu) **affiliate links*", "3d_model": "https://cdn.thingiverse.com/assets/49/2e/e5/2b/26/Gimball_Test_Print.stl"}	[{"name": "air", "provider": "thingiverse"}, {"name": "Demo", "provider": "thingiverse"}, {"name": "Gimbal", "provider": "thingiverse"}, {"name": "Spinner", "provider": "thingiverse"}, {"name": "Test", "provider": "thingiverse"}, {"name": "Useless", "provider": "thingiverse"}]	f	thingiverse	thingiverse	provider_api
+\N	https://www.thingiverse.com/thing:2823006	https://cdn.thingiverse.com/renders/ed/47/66/f7/43/3dee6caf0e321f796b4d34b78945a4ee_display_large.jpg	https://cdn.thingiverse.com/renders/ed/47/66/f7/43/3dee6caf0e321f796b4d34b78945a4ee_display_medium.jpg	\N	\N	\N	cc0	1.0	Walter Hsiao	https://www.thingiverse.com/walter	Air Spinner	{"description": "*Update 2018-04-21: Added two additional versions* I had no idea what this was for when I uploaded it. I didn't think it would be useful for anything other than as a 3D print demo, I just made it to satisfy my curiosity. But if you check out comments below, [guttyr](https://www.thingiverse.com/guttyr/about) discovered you can spin it up by blowing on it, and it can be used by speech therapists to help special needs kids get better at breath support and control. Different parts of the spinner will spin depending where you blow on it, I'm still figuring which locations and angles work best. Maybe it'll help me too as I'm frequently out of breath when I'm at altitude or underwater. It can be printed with a 0.5mm extrusion width and 2 perimeters without infill. Printed in [Atomic Filament Granite Grey PLA](https://atomicfilament.com/products/gray-granite-pla-filament) and [Octave Precision Gold PLA*](http://amzn.to/2p7JPU0) New versions printed in [Atomic Filament Starry Night PLA](https://atomicfilament.com/collections/opaque-pla-filaments-1/products/translucent-starry-night-pla) and [3D Solutech Transparent Blue PETG*](https://amzn.to/2K3migu) **affiliate links*", "3d_model": "https://cdn.thingiverse.com/assets/c7/e9/4a/9f/6b/Air_Spinner_2_-_Solid.stl"}	[{"name": "air", "provider": "thingiverse"}, {"name": "Demo", "provider": "thingiverse"}, {"name": "Gimbal", "provider": "thingiverse"}, {"name": "Spinner", "provider": "thingiverse"}, {"name": "Test", "provider": "thingiverse"}, {"name": "Useless", "provider": "thingiverse"}]	f	thingiverse	thingiverse	provider_api
