+++ openverse_catalog/dags/util/test_resources/multi_prov.tsv
@@ -1,2 +1,2 @@
-one	https://example.com/landing1	https://example.com/image1	https://example.com/thumbnail1	\N	by	4.0	alice	https://example.com/alice	title_one	{"pub_date": "1557512425", "description": "Airport", "license_url": "https://creativecommons.org/licenses/by/4.0/"}	[{"name": "travel", "provider": "test_provider"}, {"name": "flying", "provider": "test_provider"}]	f	test_provider	alice_official	provider_api	1000	500
-two	https://example.com/landing2	https://example.com/image2	https://example.com/thumbnail2	\N	by-nc	4.0	bob	https://example.com/bob	title_two	{"description": "Train", "license_url": "https://creativecommons.org/licenses/by-nc/4.0/", "raw_license_url": "https://creativecommons.org/licenses/by-nc/4.0/"}	[{"name": "travel", "provider": "next_provider"}, {"name": "rail", "provider": "other_provider"}]	f	next_provider	next_provider	provider_api	1000	500
+one	https://example.com/landing1	https://example.com/image1	https://example.com/thumbnail1	1000	500	\N	by	4.0	alice	https://example.com/alice	title_one	{"pub_date": "1557512425", "description": "Airport", "license_url": "https://creativecommons.org/licenses/by/4.0/"}	[{"name": "travel", "provider": "test_provider"}, {"name": "flying", "provider": "test_provider"}]	f	test_provider	alice_official	provider_api
+two	https://example.com/landing2	https://example.com/image2	https://example.com/thumbnail2	1000	500	\N	by-nc	4.0	bob	https://example.com/bob	title_two	{"description": "Train", "license_url": "https://creativecommons.org/licenses/by-nc/4.0/", "raw_license_url": "https://creativecommons.org/licenses/by-nc/4.0/"}	[{"name": "travel", "provider": "next_provider"}, {"name": "rail", "provider": "other_provider"}]	f	next_provider	next_provider	provider_api
