+++ openverse_catalog/dags/util/tsv_cleaner.py
@@ -3,7 +3,7 @@ import logging
 import os
 
 from common.licenses.licenses import get_license_info
-from common.storage import image
+from storage import image
 
 
 logger = logging.getLogger(__name__)
